"New document" opens now a new window (and bug)
This commit led to a new bug due to the menu manager. Since the menu is common for the app, each new window created redefines the menu to target itself, so menu is not working correctly in the other windows. To fix this bug we could detect focused windows in each command, but we will still encounter performance issues, so the best solution would be (probably) refactoring the app to manage init and Abricotine in the main process.