AbrSpellchecker
Add AbrSpellchecker to handle the dictionary loading during abrDoc creation and the switching between dictionaries. The main benefit is to allow to switch between dictionaries without reloading everything (see "test" command).

But this improvement brings a new issue: Typo() object instance is too slow to construct and it slows down the whole browser process (even if async). This is obvious when loading FR or DE dictionaries: the UI freezes during more than 10 sec.

We need to separate the dictionaries loading into another process: Webworker or main process. To be continued...