Avoid default behavior conflict on OS X
Expected behavior for an OS X Application is for Cmd+Q to quit the app. As-is, this key combination inserts a quote, not only overriding the expected behavior, but affecting the format of the document unexpectedly. Additionally, adding Shift to this key combo causes it to fall more in line with the pattern of the other format options in the group.